requirements.txt
Flask>=2.0
pymupdf>=1.23
python-docx>=0.8.11
python-magic>=0.4.27
